requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
psycopg2-binary>=2.9
gunicorn>=21.0
python-decouple>=3.8
whitenoise>=6.6
dj-database-url>=2.1
